package validation code
